Fix mobile index links
